Removed and cleaned up code/files
Also fixed search controller since there was an issue with calling the api too many times after we returned 0 results
