--- SEDaily-IOS/NotificationCenterExtension.swift
@@ -1,13 +1,0 @@
-//
-//  NotificationCenterExtension.swift
-//  SEDaily-IOS
-//
-//  Created by Craig Holliday on 6/30/17.
-//  Copyright © 2017 Koala Tea. All rights reserved.
-//
-
-import UIKit
-
-extension Notification.Name {
-    static let playingAudio = Notification.Name("playingAudio")
-}
