+++ SEDaily-IOS/Podcast.swift
@@ -108,6 +108,7 @@ extension Podcast {
     }
 }
 
+// Extension to go Encodable -> Dictionary
 extension Encodable {
     var dictionary: [String: Any] {
         return (try? JSONSerialization.jsonObject(with: JSONEncoder().encode(self))) as? [String: Any] ?? [:]
@@ -117,107 +118,3 @@ extension Encodable {
     }
 }
 
-public struct PodcastViewModel: Codable {
-    let _id: String
-    let uploadDateiso8601: String
-    let postLinkURL: URL?
-    let categories: [Int]?
-    var categoriesAsString: String {
-        get {
-            guard let categories = self.categories else { return "" }
-            let stringArray = categories.map { String(describing: $0) }
-            return stringArray.joined(separator: " ")
-        }
-    }
-    let tags: [Int]?
-    var tagsAsString: String {
-        get {
-            guard let tags = self.tags else { return "" }
-            let stringArray = tags.map { String(describing: $0) }
-            return stringArray.joined(separator: " ")
-        }
-    }
-    let mp3URL: URL?
-    let featuredImageURL: URL?
-    private let encodedPodcastTitle: String
-    private let encodedPodcastDescription: String
-    let score: Int
-    var isUpvoted: Bool = false
-    var isDownvoted: Bool = false
-    
-    var podcastTitle: String {
-        get {
-            return encodedPodcastTitle.htmlDecoded
-        }
-    }
-    
-    init(podcast: Podcast) {
-        self._id = podcast._id
-        self.uploadDateiso8601 = podcast.date
-        self.postLinkURL = URL(string: podcast.link)
-        self.categories = podcast.categories
-        self.tags = podcast.tags
-        self.mp3URL = URL(string: podcast.mp3)
-        self.featuredImageURL = URL(string: podcast.featuredImage ?? "")
-        self.encodedPodcastTitle = podcast.title.rendered
-        self.encodedPodcastDescription = podcast.content.rendered
-        self.score = podcast.score
-    }
-    
-    init() {
-        self._id = ""
-        self.uploadDateiso8601 = ""
-        self.postLinkURL = nil
-        self.categories = []
-        self.tags = []
-        self.mp3URL = nil
-        self.featuredImageURL = nil
-        self.encodedPodcastTitle = ""
-        self.encodedPodcastDescription = ""
-        self.score = 0
-    }
-}
-
-extension PodcastViewModel: Equatable {
-    public static func ==(lhs: PodcastViewModel, rhs: PodcastViewModel) -> Bool {
-        return lhs._id == rhs._id &&
-            lhs.uploadDateiso8601 == rhs.uploadDateiso8601 &&
-            lhs.postLinkURL == rhs.postLinkURL &&
-            lhs.categories ?? [] == rhs.categories ?? [] &&
-            lhs.tags ?? [] == rhs.tags ?? [] &&
-            lhs.mp3URL == rhs.mp3URL &&
-            lhs.featuredImageURL == rhs.featuredImageURL &&
-            lhs.encodedPodcastTitle == rhs.encodedPodcastTitle &&
-            lhs.encodedPodcastDescription == rhs.encodedPodcastDescription &&
-            lhs.score == rhs.score
-    }
-}
-
-extension PodcastViewModel {
-    func getLastUpdatedAsDateWith(completion: @escaping (Date?) -> Void) {
-        DispatchQueue.global().async {
-            // slow calculations performed here
-            let date = Date(iso8601String: self.uploadDateiso8601)
-            DispatchQueue.main.async {
-                completion(date)
-            }
-        }
-    }
-    
-    // This is too slow for a cell collection view call
-    func getLastUpdatedAsDate() -> Date? {
-        return Date(iso8601String: self.uploadDateiso8601)
-    }
-}
-
-extension PodcastViewModel {
-    func getHTMLDecodedDescription(completion: @escaping (String) -> Void) {
-        DispatchQueue.global().async {
-            // slow calculations performed here
-            let decodedString = self.encodedPodcastDescription.htmlDecodedWithSomeEntities ?? ""
-            DispatchQueue.main.async {
-                completion(decodedString)
-            }
-        }
-    }
-}
